// src/store/favorites/favorites.types.ts
export enum FAVORITES_ACTION_TYPES {
   FETCH_FAVORITES_START = 'favorites/FETCH_FAVORITES_START',
   FETCH_FAVORITES_SUCCESS = 'favorites/FETCH_FAVORITES_SUCCESS',
   FETCH_FAVORITES_FAILED = 'favorites/FETCH_FAVORITES_FAILED',
   REFRESH_FAVORITES_START = 'favorites/REFRESH_FAVORITES_START',
   REFRESH_FAVORITES_SUCCESS = 'favorites/REFRESH_FAVORITES_SUCCESS',
   REFRESH_FAVORITES_FAILED = 'favorites/REFRESH_FAVORITES_FAILED',
}

export enum ADD_RECIPE_TO_FAVORITES_ACTION_TYPES {
   ADD_RECIPE_TO_FAVORITES_START = 'favorites/ADD_RECIPE_TO_FAVORITES_START',
   ADD_RECIPE_TO_FAVORITES_SUCCESS = 'favorites/ADD_RECIPE_TO_FAVORITES_SUCCESS',
   ADD_RECIPE_TO_FAVORITES_FAILED = 'favorites/ADD_RECIPE_TO_FAVORITES_FAILED',
}

export enum REMOVE_RECIPE_FROM_FAVORITES_ACTION_TYPES {
   REMOVE_RECIPE_FROM_FAVORITES_START = 'favorites/REMOVE_RECIPE_FROM_FAVORITES_START',
   REMOVE_RECIPE_FROM_FAVORITES_SUCCESS = 'favorites/REMOVE_RECIPE_FROM_FAVORITES_SUCCESS',
   REMOVE_RECIPE_FROM_FAVORITES_FAILED = 'favorites/REMOVE_RECIPE_FROM_FAVORITES_FAILED',
}

export enum ADD_RESTAURANT_TO_FAVORITES_ACTION_TYPES {
   ADD_RESTAURANT_TO_FAVORITES_START = 'favorites/ADD_RESTAURANT_TO_FAVORITES_START',
   ADD_RESTAURANT_TO_FAVORITES_SUCCESS = 'favorites/ADD_RESTAURANT_TO_FAVORITES_SUCCESS',
   ADD_RESTAURANT_TO_FAVORITES_FAILED = 'favorites/ADD_RESTAURANT_TO_FAVORITES_FAILED',
}

export enum REMOVE_RESTAURANT_FROM_FAVORITES_ACTION_TYPES {
   REMOVE_RESTAURANT_FROM_FAVORITES_START = 'favorites/REMOVE_RESTAURANT_FROM_FAVORITES_START',
   REMOVE_RESTAURANT_FROM_FAVORITES_SUCCESS = 'favorites/REMOVE_RESTAURANT_FROM_FAVORITES_SUCCESS',
   REMOVE_RESTAURANT_FROM_FAVORITES_FAILED = 'favorites/REMOVE_RESTAURANT_FROM_FAVORITES_FAILED',
}